# docs/swagger.yaml
definitions:
  response.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
  /book/detail/{bookId}:
    get:
      consumes:
      - application/json
      description: 书详情
      parameters:
      - description: "1"
        in: path
        name: bookId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            x-token:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
  /book/list:
    get:
      consumes:
      - application/json
      description: 获取书籍列表
      parameters:
      - description: "1"
        in: query
        name: pageNum
        required: true
        type: integer
      - description: "10"
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            x-token:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
  /book/query:
    get:
      consumes:
      - application/json
      description: 搜索书
      parameters:
      - description: 三体
        in: query
        name: query
        required: true
        type: string
      - description: "1"
        in: query
        name: pageNum
        required: true
        type: integer
      - description: "10"
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            x-token:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
  /section/list:
    get:
      consumes:
      - application/json
      description: 章节列表
      parameters:
      - description: "1"
        in: query
        name: pageNum
        required: true
        type: integer
      - description: "10"
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            x-token:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
swagger: "2.0"